CRDL-309 Store the snapshotversion as a Long: it's an arbitrary-precision number in Denodo

+++ app/uk/gov/hmrc/crdlcache/repositories/LastUpdatedRepository.scala
@@ -52,7 +52,7 @@ class LastUpdatedRepository @Inject() (val mongoComponent: MongoComponent)(using
   def setLastUpdated(
     session: ClientSession,
     codeListCode: CodeListCode,
-    snapshotVersion: Int,
+    snapshotVersion: Long,
     lastUpdated: Instant
   ): Future[Unit] = {
     collection
